Renamed all libs from esp8266-xxx to esp-xxx
