+++ platformio.ini
@@ -30,7 +30,7 @@ build_src_filter =
   +<Common/channel_order.cpp>
   +<Common/diversity.cpp>
   +<Common/fhss.cpp>
-  +<Common/esp/esp8266.c>
+  +<Common/esp/esp.c>
   +<Common/libs/filters.cpp>
 
 [env_common_esp32]
@@ -46,7 +46,7 @@ build_src_filter =
   +<Common/channel_order.cpp>
   +<Common/diversity.cpp>
   +<Common/fhss.cpp>
-  +<Common/esp/esp8266.c>
+  +<Common/esp/esp.c>
   +<Common/libs/filters.cpp>
 
 [env:rx-dev-900]
@@ -94,4 +94,4 @@ build_src_filter =
   ${env_common_esp82xx.build_src_filter} 
   +<modules/sx12xx-lib/src/sx128x.cpp>
 build_flags = 
-  -DRX_GENERIC_2400_PA_D
+  -DRX_GENERIC_2400_PA_TD
